Deep Learning - earthquake detection using DAS data and CNN model
Fibers optic Distributed Acoustic Sensing (DAS) has rapidly progressed from lab scale to real-life applications. It was intended to be as a new form of perimeter security sensor. Currently it is used in other application such as pipelines’ protection, as part of monitoring flow and seismic data in oil and gas wells, information system for traffic flow especially on highways and monitoring condition of track and train on railroad